requirements.txt
streamlit
tensorflow
scikit-learn
pandas
numpy
matplotlib
Pillow
gdown
opencv-python
